server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
aiofiles>=23.2.0
pydantic>=2.0.0
